Update ISSUE_TEMPLATE.md
I'm not a fan of gist or pastebin as it produces a lot of search machine results without any context, annoying when one wants to find solutions for a certain error message. I hence prefer having bug reports, discussion, context with the logs combined on a single page. On GitHub, via markdown details, it can be hidden behind an expansion button.