Do not assign obsolete S.M.A.R.T. attribute
With newer Python versions, this causes an "object has no attribute 'supports_smart'", since pySMART does not provide it. "smart_capable" and "smart_enabled" do exist instead, but currently we do not use it anyway. The value is hence removed completely until someone finds time to re-implement it in a reasonable way, e.g. to enhance error messages when S.M.A.R.T. does not work.

Additionally, the still valid "is_ssd" is removed from the array, as we do not use it. We shall only handle values which we do actually use at some point.

Signed-off-by: MichaIng <micha@dietpi.com>