Update systemd unit and make it generic (#38)
In systemd it is not necessary to fork a process into background which implies the need to use Type=forking and create a PID file to allow systemd tracking its state. Instead one can run any process in foreground and use default Type=simple which implies STDOUT to systemd journal.

Defining the group in a systemd service has the downside that supplementary group memberships of the users are lost. It is hence usually advised to define the user only, in which case its primary group and supplementary group memberships are effective, as one might expect.

In the systemd journal, logs appear prefixed with the main executable by default, which is "python3" here. The system log identifier is hence set to "HTPC Manager" to allow identifying and distinguishing them better.

HTPC Manager has been tested to work fine with Python optimisation options. It reduces the size of the .pyc cache files and hence memory usage a little.

Since this systemd unit does not use any fancy new options, it can be assumed to work on any distribution. The used paths follow FHS and the usual convention for manually installed software, to avoid conflicts with package managers and build scripts.

Signed-off-by: MichaIng <micha@dietpi.com>